+++ online-courses-springboot-app/src/main/java/com/online/courses/dao/ReviewDao.java
@@ -5,5 +5,5 @@ import java.util.List;
 import com.online.courses.dto.ReviewFormBean;
 
 public interface ReviewDao {
-	public List<ReviewFormBean> fetchStudentReviews (int studentId) throws Exception;
+	public List<ReviewFormBean> fetchStudentReviews (long studentId) throws Exception;
 }
