+++ online-courses-springboot-app/src/main/java/com/online/courses/repo/StudentMstRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.online.courses.models.StudentMst;
 
 @Repository
-public interface StudentMstRepo extends JpaRepository<StudentMst, Integer> {
+public interface StudentMstRepo extends JpaRepository<StudentMst, Long> {
 
 }
